requirements.txt
flask
rembg
pillow
onnxruntime
python-dotenv
numpy
scipy